+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.10.0 - 2015-08-18
 
+- Change interface to allow for custom registries.
 - Rename the `class` metadata property to `classes`. The convenience property
   is also now called `classes`, e.g. `element.classes.contains('abc')`.